Multi-tenant support
With one user-pool. Using `identityClaim('custom:tenant_id')` to make a tenant owner of data records and allowing any user who is member of the tenant owner to access those data records.

For authentication the ID token has been configured because it contains the custom attribute "custom:tenant_id" which is required for authorization and setting the owner automatically when a record is created. It seems secure to use the ID token because it seems to also be cryptographically signed.  The documentation [here](https://docs.aws.amazon.com/cognito/latest/developerguide/amazon-cognito-user-pools-using-the-id-token.html) also says "The ID token can also be used to authenticate users to your resource servers or server applications".